backend/requirements.txt
asn1crypto==1.1.0
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-heroku==0.3.1
django-modeltranslation==0.13.3
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-storages==1.7.2
djangorestframework==3.9.4
dropbox==9.4.0
gunicorn==19.9.0
idna==2.8
isort==4.3.21
jsonfield==2.0.2
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.17.1
oauthlib==3.0.1
passlib==1.7.1
pep8==1.7.1
Pillow==6.1.0
psycopg2==2.8.3
pudb==2019.1
pycodestyle==2.5.0
pycparser==2.19
pyDOE==0.3.8
Pygments==2.4.2
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
scipy==1.3.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
urwid==2.0.1
whitenoise==4.1.3
wrapt==1.11.2
